refactor: argument order
